Refact: refatorando todo o codigo para utilizar o id de usuario que agora e salvo no token

+++ backend/entitys/creator.go
@@ -3,7 +3,7 @@ package entitys
 type Creator struct {
 	ID       int `gorm:"autoIncrement;primaryKey"`
 	Name     string
-	LeftOver float64
+	LeftOver uint64
 	UserID   int
 	User     User
 }
